[dev] implement user create

+++ src/main/java/mystream/user/service/UserService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
+import mystream.user.dto.NewStreamDto;
 import mystream.user.dto.SignUpDto;
 import mystream.user.dto.UserDto;
 import mystream.user.entity.Email;
@@ -16,9 +17,7 @@ import mystream.user.exceptions.InvalidUserProfileException;
 import mystream.user.exceptions.NotFoundException;
 import mystream.user.repository.UserRepository;
 import mystream.user.service.external.BroadcastServiceClient;
-import mystream.user.service.external.BroadcastServiceProducer;
-import mystream.user.service.external.ChannelServiceClient;
-import mystream.user.service.external.event.NewStreamEvent;
+import mystream.user.utils.ApiResponse.ApiResult;
 
 @Service
 @Transactional
@@ -29,13 +28,17 @@ public class UserService {
   private final PasswordEncoder passwordEncoder;
 
   private final BroadcastServiceClient broadcastServiceClient;
-  private final ChannelServiceClient channelServiceClient;
-
-  private final BroadcastServiceProducer broadcastServiceProducer;
 
   public UserDto findUserById(Long id) {
     return userRepository.findUserDtoByIdTo(id)
       .orElseThrow(() -> new NotFoundException("not found user. " + id));
+  }
+  
+  public void delete(Long id) {
+    Optional<User> user = userRepository.findById(id);
+    if (user.isPresent()) {
+      userRepository.delete(user.get());
+    }
   }
 
   public UserDto create(SignUpDto signUpDto) {
@@ -57,23 +60,13 @@ public class UserService {
 
     User saved = userRepository.save(user);
     
-    // // reqquest create new stream
-    NewStreamEvent newStreamEvent = new NewStreamEvent(saved.getId(), saved.getUsername());
-
-    try {
-      broadcastServiceProducer.createStream(newStreamEvent);
-    } catch (RuntimeException e) {
+    // reqquest create new stream
+    NewStreamDto newStreamDto = new NewStreamDto(saved.getId(), saved.getUsername());
+    ApiResult<?> result = broadcastServiceClient.createStream(newStreamDto);
+    if (!result.isSuccess()) {
       userRepository.delete(saved);
-      throw new RuntimeException("internal system error");
+      throw new InvalidSignupException("stream create fail");
     }
-
-    // // reqquest create new channel
-    // NewChannelDto newChannelDto = new NewChannelDto(saved.getId());
-    // result = channelServiceClient.createChannel(newChannelDto);
-    // if (!result.isSuccess()) {
-    //   userRepository.delete(saved);
-    //   throw new InvalidSignupException("channel create faile");
-    // }
 
     return new UserDto(saved);
   }
